requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.11


